+++ NEWS.md
@@ -2,6 +2,7 @@
 
 * Features:
     * Incorporate `alphavantager`, a lightweight API to the [Alpha Vantage financial data provider](https://www.alphavantage.co/).
+    * Integrate `Rblpapi`, R interface to "Bloomberg"
 
 ## tidyquant 0.5.3